backend/pyproject.toml
[project]
name = "be-vexo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "keras>=3.10.0",
    "numpy>=2.1.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "tensorflow>=2.19.0",
    "uvicorn>=0.34.3",
    "pandas>=2.0.0",
    "openpyxl>=3.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "typing-extensions>=4.0.0",
]
